Slight Rewrite
## Features
 - Added support for (in theory) boolean keys.
 - Added "Not Supported"  as a data output if the value doesn't match a table, string, number or boolean. (For functions for instance)
 - Added "notsupported" key value for any keys that do not match a string, number or boolean.
## Performance
 - ~20% Faster overall
 - No global variables used for deserialize
## Technicals
#### Where did the performance come from?
The first if-check for the key-type will be a string, followed by a number then a boolean. This is generally faster as it will not check less likely key types first.
Additionally, for an indexed table, it now uses a normal for-loop instead of pairs. This is faster, however, it does mean data is lost entirely instead of partially if you had any keys on the same level. 

No longer uses string.format and instead uses good old string concatenation. It is slightly faster than splitting it up into a table. 
Made deserialize a single line and avoided creating a global variable.

The new code-flow scheme now avoids a bit of function call overhead by only calling a new function when a table value appears.